Move CassandraDocker to testkit
