+++ build.sbt
@@ -270,6 +270,7 @@ lazy val cli = project
 
 lazy val sourcing = project
   .in(file("sourcing"))
+  .dependsOn(testkit % "test->compile")
   .settings(name := "sourcing", moduleName := "sourcing")
   .settings(shared, compilation, assertJavaVersion, coverage, release)
   .settings(
