+++ sourcing/src/test/scala/ch/epfl/bluebrain/nexus/sourcing/projections/ProjectionsSpec.scala
@@ -8,6 +8,7 @@ import cats.effect.{ContextShift, IO, Resource, Timer}
 import cats.implicits._
 import ch.epfl.bluebrain.nexus.sourcing.projections.ProjectionProgress.{NoProgress, OffsetProgress}
 import ch.epfl.bluebrain.nexus.sourcing.projections.ProjectionsSpec.SomeEvent
+import ch.epfl.bluebrain.nexus.testkit.cassandra.{CassandraDocker, CassandraDockerModule}
 import com.typesafe.config.ConfigFactory
 import distage.{DIKey, ModuleDef}
 import io.circe.{Decoder, Encoder}
